enhanced recipe ownership checks(user cannot access edit recipe page if they're not the owner) and improve UI elements : fixed wrong texts / Guests cannot see recipe delete/edit buttons, Users cannot see delete buttons of recipes they didn't create

+++ public/scripts/recipes-details.js
@@ -2,43 +2,45 @@ import { createModal } from "./modal.js";
 import { createReviewModal } from "./reviewModal.js";
 const trashcan = document.querySelector(".delete-btn");
 
-
-trashcan.addEventListener("click", (e) => {
-  e.preventDefault();
-  createModal(
-    "Are you sure?",
-    "Do you really want to delete this recipe? This action cannot be undone.",
-    async() => {
-      try {
-      const endpoint = `/recipes/${trashcan.dataset.doc}`;
-      const response = await fetch(endpoint, {
-        method: "DELETE",
-      });
-      const modal = document.querySelector('.modal-backdrop')
-        if(response.ok){
-          setTimeout(() => {
-            modal.innerHTML = `
-                    <div class="modal-container">
-                        <div style="display: flex; justify-content: center;">
-                      <svg class="checkmark" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 52 52"> <circle class="checkmark__circle" cx="26" cy="26" r="25" fill="none"/> <path class="checkmark__check" fill="none" d="M14.1 27.2l7.1 7.2 16.7-16.8"/>
-                    </svg>
-                    </div>
-                      `;
-          }, 1000);
-          setTimeout(() => {
-            window.location.href = '/recipes'
-          }, 2000);
-        }else{
-          modal.remove();
-          const errorData = await response.json();
-          showNotification(errorData.error,'var(--error-color)');
+if(trashcan){
+  trashcan.addEventListener("click", (e) => {
+    e.preventDefault();
+    createModal(
+      "Are you sure?",
+      "Do you really want to delete this recipe? This action cannot be undone.",
+      async() => {
+        try {
+        const endpoint = `/recipes/${trashcan.dataset.doc}`;
+        const response = await fetch(endpoint, {
+          method: "DELETE",
+        });
+        const modal = document.querySelector('.modal-backdrop')
+          if(response.ok){
+            setTimeout(() => {
+              modal.innerHTML = `
+                      <div class="modal-container">
+                          <div style="display: flex; justify-content: center;">
+                        <svg class="checkmark" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 52 52"> <circle class="checkmark__circle" cx="26" cy="26" r="25" fill="none"/> <path class="checkmark__check" fill="none" d="M14.1 27.2l7.1 7.2 16.7-16.8"/>
+                      </svg>
+                      </div>
+                        `;
+            }, 1000);
+            setTimeout(() => {
+              window.location.href = '/recipes'
+            }, 2000);
+          }else{
+            modal.remove();
+            const errorData = await response.json();
+            showNotification(errorData.error,'var(--error-color)');
+          }
+        } catch (error) {
+          showNotification(error,'var(--error-color)');
         }
-      } catch (error) {
-        showNotification(error,'var(--error-color)');
       }
-    }
-  );
-});
+    );
+  });
+}
+
 
 // delete review
 const deleteRecipeBtn=document.querySelector('.deleteRecipeBtn');
@@ -201,5 +203,12 @@ reviewRecipeBtns.forEach(btn => {
   });
 });
 
-
-
+      //if user tries to enter the edit page of a recipe he doesn't own we'll redirect him to the recipe with an error
+      document.addEventListener('DOMContentLoaded', function() {
+        const urlParams = new URLSearchParams(window.location.search);
+        const error = urlParams.get('error');
+        
+        if (error === 'unauthorized') {
+            showNotification("You don't have permission to edit this recipe.", 'var(--error-color)');
+        }
+      });
